TVG-103 Move reminder requests into new file
