+++ frontend/src/requests/index.ts
@@ -5,13 +5,7 @@ import {
     CurrentUser,
     Guide,
     LoginData,
-    NewShowPayload,
     NewUserDetails,
-    Reminder,
-    ReminderFormValues,
-    SearchItemPayload,
-    SearchItem,
-    ShowData,
     ShowEpisode,
     User,
 } from "../utils/types";
@@ -22,27 +16,6 @@ const getGuide = async () => {
 
 
 
-const getReminders = async () => {
-    const reminders = await getRequest<Reminder[]>("/reminders");
-    return reminders;
-};
-const addReminder = async (reminder: ReminderFormValues, token: string) => {
-    return await postRequest<ReminderFormValues, Reminder>(
-        `/reminders`,
-        reminder,
-        { Authorization: `Bearer ${token}` }
-    );
-};
-const editReminder = async (reminderDetails: ReminderFormValues, token: string) => {
-    return await putRequest<ReminderFormValues, Reminder>(
-        `/reminder/${reminderDetails.show}`,
-        reminderDetails,
-        { Authorization: `Bearer ${token}` }
-    );
-}
-const deleteReminder = async (show: string, token: string) => {
-    return await deleteRequest(`/reminder/${show}`, { Authorization: `Bearer ${token}` });
-};
 
 const updateShowEpisode = async (episode: ShowEpisode, token: string) => {
     return await putRequest<ShowEpisode, ShowEpisode>(
@@ -98,15 +71,12 @@ const login = async (loginDetails: LoginData) => {
     return currentUser;
 };
 
+export * from "./reminders";
 export * from "./shows";
 export * from "./search-items";
 
 export {
     getGuide,
-    getReminders,
-    addReminder,
-    editReminder,
-    deleteReminder,
     updateShowEpisode,
     getUser,
     registerNewUser,
