+++ frontend/src/requests/reminders.ts
@@ -1,0 +1,25 @@
+
+import { deleteRequest, getRequest, postRequest, putRequest } from "./api-client";
+import { Reminder, ReminderFormValues } from "../utils/types";
+
+const getReminders = async () => {
+    const reminders = await getRequest<Reminder[]>("/reminders");
+    return reminders;
+};
+const addReminder = async (reminder: ReminderFormValues, token: string) => {
+    return await postRequest<ReminderFormValues, Reminder>(
+        `/reminders`,
+        reminder,
+        { Authorization: `Bearer ${token}` }
+    );
+};
+const editReminder = async (reminderDetails: ReminderFormValues, token: string) => {
+    return await putRequest<ReminderFormValues, Reminder>(
+        `/reminder/${reminderDetails.show}`,
+        reminderDetails,
+        { Authorization: `Bearer ${token}` }
+    );
+}
+const deleteReminder = async (show: string, token: string) => {
+    return await deleteRequest(`/reminder/${show}`, { Authorization: `Bearer ${token}` });
+};
